Move logic to separate functions.
