+++ packages/generator/src/generators/shortcodes/generateGitHub.ts
@@ -1,0 +1,31 @@
+import fetchAndCache from '../../loaders/fetchAndCache';
+import writeDataset from '../../helpers/writeDataset';
+import { ShortcodeDataMap } from '../../types';
+
+export default async function generateGitHub() {
+  const shortcodes: ShortcodeDataMap = {};
+  const response = await fetchAndCache<Record<string, string>>(
+    'https://api.github.com/emojis',
+    'temp/github-emojis.json',
+    (text) => JSON.parse(text),
+    {
+      headers: {
+        Accept: 'application/vnd.github.v3+json',
+        'User-Agent': 'Emojibase',
+      },
+    },
+  );
+
+  Object.entries(response).forEach(([code, url]) => {
+    const match = url.match(/emoji\/unicode\/([\da-z-]+)\.png/i);
+
+    if (match) {
+      shortcodes[match[1].toUpperCase()] = code;
+    }
+  });
+
+  await Promise.all([
+    writeDataset(`en/shortcodes/github.raw.json`, shortcodes),
+    writeDataset(`en/shortcodes/github.json`, shortcodes, true),
+  ]);
+}
